Update README.md
Introducing the ChatViewer plugin for Minecraft 1.19.3! This plugin offers a unique solution for players who want to keep an eye on their server's chat, even when they're unable to join the game. With ChatViewer, you can use a separate desktop application to view the chat in real-time, providing an added layer of security and convenience for those who value their privacy and need to stay connected to their community.

The ChatViewer plugin is easy to set up and use. Simply install the plugin on your Minecraft server, and then download the accompanying desktop application. Once you've connected the two, you'll be able to see all the chat happening on your server, even when you're away from the game.

ChatViewer is perfect for players who want to stay connected to their community, but can't always be in the game. It's also great for server administrators who want to keep an eye on their server's chat, even when they're away from the game. So don't wait, download ChatViewer today and experience the convenience and security of an external chat viewer for your Minecraft server!